% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-package.R
\docType{package}
\name{binary-package}
\alias{binary}
\alias{binary-package}
\title{binary: Exact tests for two proportions}
\description{
Exact tests for superiority and non-inferiority of two proportions,
    power and sample size calculations.
}
\keyword{internal}
