% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_loop.R
\name{classifier_loop}
\alias{classifier_loop}
\title{Title}
\usage{
classifier_loop(Xtrain.feather, y.feather, Xtest.feather, data_checks_func,
  train_func, iter, pct_train)
}
\arguments{
\item{pct_train}{}
}
\description{
Title
}

