% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_checks_binary.R
\name{data_checks_binary}
\alias{data_checks_binary}
\title{Title}
\usage{
data_checks_binary(Xtrain, Xtest, y)
}
\arguments{
\item{y}{}
}
\value{
TRUE boolen indicating success
}
\description{
Title
}

