% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-functions.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
Function call rhs(lhs).
}
\description{
Define magrittr's Pipe operator "%>%" locally for simplified programming.
}
