% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_OUPUT.R
\name{init_OUPUT}
\alias{init_OUPUT}
\title{init_OUPUT}
\usage{
init_OUPUT(n_train_full, n_test_full, iter, pct_train)
}
\arguments{
\item{n_train_full}{number of training examples}

\item{n_test_full}{number of test examples}

\item{iter}{how many iterations to run}

\item{pct_train}{what PCT 0,1 to train on)}
}
\value{
list PCV cv predictions
}
\description{
init_OUPUT
}
\examples{
OUPUT <- init_OUPUT(1000,1000,100,.25)


}

