% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{prob.derivative}
\alias{prob.derivative}
\title{Calculate derivative of rejection probability}
\usage{
prob.derivative(p_C.vec, x_E, x_C, n_E, n_C, method, delta)
}
\arguments{
\item{p_C.vec}{data frame with variables \mjseqn{x_E} and \mjseqn{x_C}}

\item{x_E}{vector of values belonging to region}

\item{x_C}{vector of values belonging to region}

\item{n_E}{Sample size in experimental group.}

\item{n_C}{Sample size in control group.}

\item{method}{Specifies the effect measure/test statistic. One of "RD", "RR", or "OR".}

\item{delta}{Non-inferiority margin.}
}
\value{
Vector of values of the derivative
}
\description{
Calculate derivative after \mjseqn{p_E} of probability of a specific region under \mjseqn{H_0}.
}
