% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_Inferiority_RD_RR.R
\name{samplesize_appr}
\alias{samplesize_appr}
\title{TODO}
\usage{
samplesize_appr(p_EA, p_CA, delta, alpha, beta, r, method)
}
\arguments{
\item{p_EA}{assumed response rate in the experimental group under the alternative}

\item{p_CA}{assumed response rate in the control group under the alternative}

\item{delta}{TODO}

\item{alpha}{maximal type I error rate}

\item{beta}{maximal type II error rate}

\item{r}{allocation ratio}

\item{method}{TODO}
}
\description{
TODO
}
