% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_Inferiority_RD_RR.R
\name{samplesize_exact}
\alias{samplesize_exact}
\title{Compute exact sample size}
\usage{
samplesize_exact(
  p_EA,
  p_CA,
  delta,
  alpha,
  beta,
  r,
  size_acc = 3,
  method,
  better
)
}
\arguments{
\item{p_EA}{assumed response rate in the experimental group under the alternative}

\item{p_CA}{assumed response rate in the control group under the alternative}

\item{delta}{TODO}

\item{alpha}{maximal type I error rate}

\item{beta}{maximal type II error rate}

\item{r}{allocation ratio}

\item{size_acc}{TODO}

\item{method}{TODO}

\item{better}{TODO}
}
\value{
Sample sizes per group (n_C, n_E), critical value and exact power.
}
\description{
Calculate exact sample size for method "X and specified
level alpha, beta, allocation ratio r = n_E/n_C, true rates p_CA, p_EA and
OR-NI-margin delta
}
\details{
Accuracy of calculating the critical value can be specified by size_acc.
}
