% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boschloo-functions.R
\name{samplesize_normal_appr}
\alias{samplesize_normal_appr}
\title{Calculate approximate sample size for normal approximation test}
\usage{
samplesize_normal_appr(p_EA, p_CA, alpha, beta, r)
}
\arguments{
\item{p_EA}{assumed response rate in the experimental group under the alternative}

\item{p_CA}{assumed response rate in the control group under the alternative}

\item{alpha}{maximal type I error rate}

\item{beta}{maximal type II error rate}

\item{r}{allocation ratio}
}
\value{
Sample sizes per group (n_C, n_E).
}
\description{
Calculate approximate sample size for normal approximation test for specified
level alpha, power, allocation ratio r = n_E/n_C and true rates p_CA, p_EA.
}
