% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Non_Inferiority_RD_RR.R
\name{test_RR}
\alias{test_RR}
\title{Calculate RR test statistic}
\usage{
test_RR(x_E, x_C, n_E, n_C, delta, better)
}
\arguments{
\item{x_E}{Vector of number of events in experimental group.}

\item{x_C}{Vector of number of events in control group.}

\item{n_E}{Sample size in experimental group.}

\item{n_C}{Sample size in control group.}

\item{delta}{Non-inferiority margin.}

\item{better}{"high" if higher values of \mjseqn{x_E} favour the alternative
hypothesis and "low" vice versa.}
}
\value{
Vector of values of the RD test statistic.
}
\description{
\code{test_RR} returns the value of the Farrington-Manning test statistic
for non-inferiority of the risk ratio between two proportions.
}
\details{

}
\examples{
test_RD(3, 4, 10, 10, 0.2, "high")
}
