% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_OUPUT.R
\name{init_OUPUT}
\alias{init_OUPUT}
\title{init_OUPUT}
\usage{
init_OUPUT(n_train_full, n_test_full, iter, pct_train)
}
\arguments{
\item{n_train_full}{number of training examples}

\item{n_test_full}{number of test examples}

\item{iter}{how many iterations to run}

\item{pct_train}{what % to train on (10-25% current use)}
}
\value{
list PCV cv predictions, PT training predictions, IDX,
    training indices, DATA meta data from training
}
\description{
init_OUPUT
}

