% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{p_value}
\alias{p_value}
\title{Calculate p-value(s)}
\usage{
p_value(
  x_E.,
  x_C.,
  n_E,
  n_C,
  method,
  delta = NULL,
  size_acc = 3,
  better = c("high", "low"),
  calc_method = c("uniroot", "grid search")
)
}
\arguments{
\item{x_E.}{Number of events in experimental group.}

\item{x_C.}{Number of events in control group.}

\item{n_E}{Sample size in experimental group.}

\item{n_C}{Sample size in control group.}

\item{method}{Specifies the effect measure/test statistic. One of "RD", "RR", or "OR".}

\item{delta}{Non-inferiority margin.}

\item{size_acc}{Accuracy of grid}

\item{better}{"high" if higher values of x_E favour the alternative 
hypothesis and "low" vice versa.}

\item{calc_method}{"grid search" or "uniroot"}
}
\value{
A list with the two elements \code{p_max} and \code{p_vec}.
\code{p_max} is the maximum p-value and most likely servers as "the one" p-value.
\code{p_vec} is a named vector. The names indicate the true proportion pairs
$(p_E, p_C)$ with $e(p_E, p_C) = \delta$ that underly the calculation of
the p-values. It can be used for plotting the p-value versus the true proportions.
}
\description{
\code{p_value} returns the vector of p-values (dependent on the true $H_0$ proportions)
and its maximum of the test for non-inferiority of two proportions specified by \code{method}.
}
\details{
If higher values of $x_E$ favour the alternative hypothesis (\code{better = "high"}), we are interested
in testing the null hypothesis
$$H_0: e(p_E, p_C) \le \delta ,$$
where $e$ is one of the effect measures risk difference (\code{method = "RD"}),
risk ratio (\code{method = "RR"}), or odds ratio (\code{method = "OR"}).
The test statistic for risk difference is
$$T_{\RD, \delta}(x_E, x_C) = \frac{\hat p_E - \hat p_C - \delta}{\sqrt{\frac{\tilde p_E(1 - \tilde p_E)}{n_E} + \frac{\tilde p_C(1 - \tilde p_C)}{n_C}}},$$
where $\tilde p_C = \tilde p_C(x_E, x_C)$ is the MLE of $p_C$ and
$\tilde p_E = \tilde p_C + \delta$ is the MLE of $p_E$ under $p_E - p_C = \delta$.
High values of $T_{\RD, \delta}$ favour the alternative hypothesis.
The test statistic for risk ratio is
$$T_{\RR, \delta}(x_E, x_C) = \frac{\hat p_E - \delta \cdot \hat p_C}{\sqrt{\frac{\tilde p_E(1 - \tilde p_E)}{n_E} + \delta^2\frac{\tilde p_C(1 - \tilde p_C)}{n_C}}},$$
where $\tilde p_C = \tilde p_C(x_E, x_C)$ is the MLE of $p_C$ and
$\tilde p_E = \tilde p_C + \delta$ is the MLE of $p_E$ under $p_E / p_C = \delta$.
High values of $T_{\RR, \delta}$ favour the alternative hypothesis.
The test statistic for Odds Ratio
$$ T_{\OR, \delta} = \sum\limits_{k = X_E}^\infty  f_{\ncHg(X_E+X_C, n_E, n_C, \delta)}(k) $$
is based on Fisher's non-central hypergeometric distribution with density
$$ f_{\ncHg(s, n_E, n_C, \delta)}(k) = \frac{\binom{n_E}{k}\cdot \binom{n_C}{s-k}\cdot \delta^k}{\sum\limits_{l \in A_{s, n_E, n_C}} \binom{n_E}{l}\cdot \binom{n_C}{s-l}\cdot \delta^l}, $$
where $A_{s, n_E, n_C} = \{\max(0, s-n_C), \dots, \min(n_E, s)\}$.
The density is zero if $k < \max(0, s-n_C)$ or $k > \min(n_E, s)$.
Small values of $T_{\OR, \delta}$ favour the alternative hypothesis.
}
\examples{
p_value(
  x_E. = 3,
  x_C. = 4,
  n_E = 10,
  n_C = 10,
  method = "RD",
  delta = -0.1,
  size_acc = 3,
  better = "high"
)
}
